backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.7
fpdf2==2.7.6
pandas==2.0.3
python-docx==0.8.11
openpyxl==3.1.2
